Fix error emitted from nuclide-diff-gutters
Summary:
Turns out if you have an observable with `.first()` then completes before it emits a value, it will throw an error. This was causing the observable to throw in some cases and cause the feature to stop working. `.take(1)` doesn't throw

Repro'd by adding a `.catch` to the end of the observable.  Opening some files, checking out a new revision, then using 'Close other tabs'.

Reviewed By: evangrayk

Differential Revision: D9813359

fbshipit-source-id: 71f37900122be68b08031309066635636732ed30