Wrap all combineEpics to handle epic errors individually
Summary:
If an epic has an uncaught error, it will bubble up to the root, and cause all epics to be resubscribed. This can break things when working with hot observables, and you end up with stale handles in some situations. The fix is to handle errors in each epic individually. This fix had already been applied to a couple of packages (where it's fixed major reliability bugs), but we should really just do this everywhere, so copy-pastes will pull from a sane template. To do so, we just encapsulate some logic that everyone seemed to copy into a function.

Some epics were not set up to catch top level errors at all. So this change will eat up those fatal errors in favor of logging something---I'm not sure if this is desirable or not. This makes sense for the packages I usually work with, but some may prefer being really noisy about errors.

Minor question: Is `combineEpicsFromImports` a good name?

Reviewed By: wbinnssmith

Differential Revision: D9818180

fbshipit-source-id: 16837efde8b1cf40e893d54f5a7ce89eac0f54ee