make error.code compatible with Nuclide's expectations
Summary:
ThriftFS returns the unix numerical error code associated with file operations (e.g. 2). Nuclide expects the actual error string (e.g. ENOENT).

This diff changes the format of ThriftFS errors to return the actual string as the error `code`, and the number as the `numericalErrorCode`.

It would be good to get this out with Monday's release so we can keep the GK on.

Reviewed By: wanderley

Differential Revision: D9841270

fbshipit-source-id: 067948299081b49efc09472ecaf62d23bce4e7d6