Java Debugger: Support Java Breakpoints in the `xplat/js` directory
Summary:
If the stack frame's path includes /js/ and we cannot find the source, then we look in the xplat/js directory.

Note: This will only work if the user has fbsource mounted. If they have both fbandroid and xplat mounted, I'm not sure that this will work.

Secondary Note: The returned path is a lowercased version of the actual directory name. This means that Nuclide will open a new tab for this file. I have seen this issue occur in other scenarios as has Herman.

Reviewed By: ebluestein

Differential Revision: D9806946

fbshipit-source-id: bfa5979d87249dc1b8d5dbd555e2d51076711bc5