Implement atom-copy and atom-paste that can behave as pbcopy and pbpaste
Summary:
This adds new messages that can be sent to connected Atom windows, specifically ones that translate to `clipboard.readText` and `clipboard.writeText`.

Once this lands, we can ship `pbcopy` and `pbpaste` as part of the nuclide scripts, so it will work on a remote server with a Nuclide connection the same way the magical `atom` command works.

Most of the environment setup is copypasted from the other tools (atom/atom-notif/nuclide-connections).

Reviewed By: semmypurewal

Differential Revision: D9687271

fbshipit-source-id: 8d99bb781c430fb5f7d292b9a6f6d0fd94099b89