Show an error box instead of crashing the debugger when unable to open stack frame.
Summary:
When attempting to open a stack frame in the debugger that doesn't have a corresponding file, we currently throw an unhandled exception which brings down the entire debugger process, necessitating a restart of the debugging session.

Instead, this change just makes us return null, which then causes a helpful error box to appear and explain that the file can't be opened.

Reviewed By: ebluestein

Differential Revision: D9927593

fbshipit-source-id: b0a4ee1ad9b569250951e5fff6ff9cf1ec033008