Debugger: Fix regression in attach/launch buttons and collapse while debugging
Summary: A regression caused the launch attach buttons to disappear from the debugger pane while debugging is in progress. These buttons are needed to add additional targets for multi-target debugging. Add them back but also put them in a section that collapses to reduce visual noise when debugging is already in progress.

Reviewed By: mostafaeweda

Differential Revision: D9838790

fbshipit-source-id: 716ecdcbd04e06c9b908edbb892d87dccb5fb86a