Create an accessible `<Link>` component
Summary:
Using bare `<a>` tags in Atom means that they are not keyboard accessible -- pressing Enter/Return while a link is focused does not open it. Adding the `native-key-bindings` class to an anchor does solve this, however it results in loss of other functionality like tabbing between inputs.

This listens for Atom's own `core:confirm` events and simulates a click on the anchor element itself, not unlike how browsers emulate clicks on anchors and buttons when enter or space is pressed respectively.

Reviewed By: matthewwithanm

Differential Revision: D9980319

fbshipit-source-id: 63bb809fdf8297fbbea0ce8752aab8e3b6df63f1