Close existing watchman connection when reconnection
Summary: If we try to reconnect to watchman, but have a partial or existing subscription, we should end that subscription before reconnecting, or else we might create an unbounded number of connections. In another diff I'm going to try to show notifications when we start a retrying loop and when the retry has succeeded, which should make this more robust to the user.

Reviewed By: aadisriram

Differential Revision: D10001196

fbshipit-source-id: 7113dddb30c1922ddc026a21a19eff215d6ab483