Fix Dropdown.js from crashing because of D9882859
Summary:
Using a wildcard import causes Babel to emit different code in order to try to comply with the ES2015's definition of a module. Specifically, it emits a `_includeModuleAsWildcard` method that seems to reconstruct the original object, but in this case, strips out the Menu constructors.

This doesn't seem to happen in Dev mode (why we didn't catch this), which hints that on-the-fly compilation is emitting different code than the code that `scripts/fb-setup` emits.

Reviewed By: kcaze

Differential Revision: D10013902

fbshipit-source-id: 0754119004e7aef217a65a9e9e933fa5819ff2ed