Call startsWith on the message, not error itself
Summary:
This is a silly mistake and unfortunately since this is inside the `catch` block which should save us, it errors out the entire detection observable (detection stops).

Ater we land this, I'll make a decision whether we need a hotfix release or not. In the last 24 hours this would affect 92 users, so we likely will.

Reviewed By: aadsm

Differential Revision: D10075725

fbshipit-source-id: 21ecc86e42f0b57a9e1029193ff8a66e6b713082