Watch pane changes to replace remote text editors
Summary: Currently when a new remote connection is setup nuclide scans through all pane items and replace editors with a read only placeholder to be replaced by an editable text editor. This unfortunately doensn't work when you manually serialize and deserialize workspace like bookshelf does. Turns out Nuclide was doing a full workspace scan anyway, so now instead of just on add project we also do this when a pane item is added. This change also removes the parameter passed to these replacer utils to not depend on a particular connection.

Reviewed By: evangrayk

Differential Revision: D10021965

fbshipit-source-id: d7c83b2aed2e73b61fa086f4901030b72b22421e