add 'adapter' command to see adapter info
Summary:
Command to dump the adapter capabilities

Also, remove capability caching from fbdbg as VsDebugSession now handles it, along with updating it when the launch or attach command has a new list.

Fix a bug in the hhvm adapter wrapper where it was sending a redundant attach response, causing the updated capabilities in the real attach response to be ignored.

Reviewed By: ebluestein

Differential Revision: D10014978

fbshipit-source-id: 7bf2f9aae77674ec116fde4cefe372e97f6fff03