Properly exit on adapter initialize failure
Summary:
If adapter initialize fails, we don't exit the debugger, because we get an adapter exited event. We used to ignore that during initialize, because we expect to get it if we're closing an old session and starting a new one (i.e. relaunch).

Tighten up the checks and only ignore the event if we know we're in the middle of disconnecting the old session.

Reviewed By: ebluestein

Differential Revision: D10126785

fbshipit-source-id: 60192adc333d09e19dbd951e53042175345baa21