do not assume symlinks are files
Summary:
The Remote File System (RFS) and new Thrift RFS (TRFS) have slightly different APIs which results in different information about symlinks being returned. The RFS code hardcodes whether a file is a symlink or not, while TRFS relies on stat/lstat data. This results in different output and behavior.

This fixes the regression where symlinks are hidden with the RFS. (TRFS works fine here because symlinks are only returned when they point to nothing and therefore are always files; RFS returns the symlink boolean along with actual stat data of what the symlink points to, so it can be a directory that is a symlink)

This diff also makes the behavior slightly more consistent in that symlinks that point to files that are missing do not appear in the file tree (prior to this diff TRFS displayed a file which was recognized as a symlink).

Reviewed By: matthewwithanm

Differential Revision: D10135797

fbshipit-source-id: 8402234443a989293b64786ff10e12165599d0d3