hardcode symlink data instead of relying on stat data
Summary:
TRFS now passes stat data on a symlink's target along with whether the file was a symlink.

If a symlink can't be followed, the symlink's lstat data is returned (stat data on the link itself). This means symlinks with missing targets or invalid targets (such as a recursive symlink yielding ELOOP like `ln -s alink alink`) will still show up in the file tree. The file tree will note them as Files, but will have the symlink flag appropriately set. The ELOOP error will not appear until the user clicks the file, which is appropriate.

This is an improvement from the current implementation for both Thrift and the current RFS in that it will display missing/invalid symlinks in the file tree rather than simply omitting them. Users may find some symlinks they didn't know they had :).

Reviewed By: semmypurewal

Differential Revision: D10137192

fbshipit-source-id: a03559baf539cf8d4d7d84bafd9bc1c44d0a64d1