Watchman health status bar
Summary:
Add in a new status bar component to show watchman's health. This hooks into the watchman health observable introduced in D10043813.

This is the starting point, there's a few things that I'm going to add in other diffs:
- First show up in grey until it's been disconnected for more than 30s-1m or so, then show in red
- Make clicking the status button refresh smartlog
- After reconnecting, show green "Watchman: Reconnected!" for a few seconds before disappearing
- Show details of connection in the tooltip, such as time since it became unhealthy.

This will be GK'd in the next diff (not done here for brevity)

Is "watchman: disconnected" too wordy?

Reviewed By: aadisriram

Differential Revision: D10114952

fbshipit-source-id: 476f4c329b4a7c15b48289e0dc539838047fd474