Move exception breakpoint filters to Process
Summary:
Currently exception breakpoint filters are on the DebuggerModel, but this is problematic when multi-target debugging beause different debug adapters offer different breakpoint filters.  Moving this to the process so the filters displayed are always for the currently focused debug target.

When not debugging, we cache and offer the filters for the last used debug adapter, so that users can set filters before re-starting the debugger (to catch exceptions thrown early in the target). This behavior is similar to what VS Code does.

Reviewed By: mostafaeweda

Differential Revision: D10148181

fbshipit-source-id: f8ce6b9e41f34ddf4a36487389f75b691d2d1cef