Don't query for `parents(smartlog())`
Summary:
When fetching the revision list, we query for `smartlog()+parents(smartlog())`. This causes us to pull in about 2 * (number of public commits), when we only need 1x. This is legacy from before rewriting the revisionList renderer. Now that we build children into a tree in a different way, we don't need to use the public commit's public parents to determine when a commit is public.

This was *probably* something that was used to determine when a public commit is part of a public non-master branch. However, that functionality is already broken, so we might as well speed things up and ask hg for a better way to query this. There's also no particularly easy way I could find to re-add that functionality using what we query now.

In my unscientific testing, a typical fetch went from 0.8s to 0.6s, which is about 25% speedup. This is a bigger win depending on how many public commits you're viewing (usually indirectly), which can be half of your commits if you pull a lot and make a lot of individual commits.

Reviewed By: aadisriram

Differential Revision: D10156518

fbshipit-source-id: 2f10319517c2cf35908856cbfb27881864b6c2ff