Stop constructing ugly error messages
Summary:
Unfortunately these errors sometimes leak to the user. While they shouldn't, it's a much worse experience when they do, because the message field contain the error stack and RPC message details.

With this patch, we still append the details about the error to the object but the `message` property is now the same as if the error happened locally and is usually a much easier to read string.

Reviewed By: wbinnssmith

Differential Revision: D10178154

fbshipit-source-id: ae6dc19a0bca525636d061054d73c1b969fe1b9a