Remove VCS status code from FileTreeNode
Summary: Removing VCS statuses from nodes means that we'll no longer have to create new nodes every time the status changes. Unfortunately, we still are (even after this commit) because the VCS status is stored in the "conf" object, and changing that results in creating a new node. That changes with D10212312, however.

Reviewed By: captbaritone

Differential Revision: D10212313

fbshipit-source-id: 5e33f66a110dc9d8c8ea175ead5269a949c74a78