Allow access to untokenized input
Summary:
The parser for commands currently just splits tokens based on whitespace, with an exception for single quoted strings.

Rewrite the parser so that it's possible to get at the tail of the command, starting where any parsed token starts. Use this for expression evaluation (and for the upcoming conditional breakpoints feature). This ensures that expressions are whitespace exact.

We may consider getting rid of the single quotes feature now, as the debugger can tell from context where whitespace matters.

Reviewed By: Goom11

Differential Revision: D10219207

fbshipit-source-id: 43aaf453c88bee44052185a08ba99ad8bbe33410