Support includes in `getBuckConfig`
Summary:
Buck recently added an include feature to buckconfig so that we could take different cell configs with similar values and consolidate them into a single source of truth. Unfortunately Nuclide parses the buckconfig directly, and since includes aren't a standard ini feature they do the wrong thing.

So instead, fake the implementation here.

NOTE: Really the correct thing to do is just call `buck audit config`, but as explained in the comment the performance of that command is too poor to do in situations where we need responsive UI elements.

(This diff also adds support for `$(config foo.bar)` syntax in buckconfig. This has been a Buck feature for a long time, but we apparently were lucky in never asking for a config that used it)

Reviewed By: stepanhruda

Differential Revision: D10202581

fbshipit-source-id: 5f9ab9e193b8b427b790f5c54e04e347378d61d8