determine supported functions dynamically
Summary:
The list of supported thrift functions was being maintained statically. This results in busywork each time a new thrift function is added, and may have typos. Determine this list once at startup and use it when calling filesystem functions.

Note that functions starting with `_` are not added to the list so if we decide to deprecate or hide functions in the future, we can just add an underscore prefix.

Reviewed By: semmypurewal

Differential Revision: D10220017

fbshipit-source-id: 8639fea8c316da0d6fbf5bf48fc6350bf9d42988