Compact commit header
Summary:
The previous header for each commit in interactive split was quite large and bulky, taking away valuable vertical space for code. This change slims it down so the commit number is on the same line as the title box and fold/unfold buttons.

EDIT:
I've updated it now, so that it also shows "1/n" instead of "Commit #1" which is a lot more terse. I moved the input box to the left, and removed the background color. "Leave Changes Uncommitted" is now on its own line (which is fine since nothing wraps)

Reviewed By: aadisriram

Differential Revision: D10225822

fbshipit-source-id: 2197a039510d983b20e7740de057985ce3ad3256