Ground work: add types to tunnel messages
Summary:
Various messages are sent over big-dig Tunnels such as `proxyCreated`, `proxyClosed`, `data`. The messages are objects that are converted to JSON, but they have no type. This can lead to errors because the code receiving the messages has no way to ensure the fields it expects to be there are guaranteed to be there.

This diff makes the type fully explicit based on the event that occurred rather than having the all the fields be optional.

This is part of a sequence of diffs to add IPC socket support.

Reviewed By: cs01

Differential Revision: D10283163

fbshipit-source-id: f2ce2271651fa7cd0199d4c40216d617e12da0a6