Java Debugger: Add Run In Terminal Support
Summary:
Added Run In Terminal support for the Java debugger.

On the Java debugger side of things, I have added the ability to send back requests with callbacks that get called when we receive a response with the correct request_id. Pretty cool, huh? I then used this to send the RunInTerminalRequest that when it gets the response, attaches to the appropriate port.

On the Nuclide side, I made the terminal emit a 'spawn' event when the process it is running has actually spawned. This is because the Java debugger cannot safely attach to that port until the command has spawned. A more accurate method would require waiting to receive output from the spawned process but it seems we don't need that solution at the moment.

Reviewed By: ebluestein

Differential Revision: D10368993

fbshipit-source-id: e44ea84fed38476196662bc3ac255abe954df9e4