Properly set up document root for hhvm on on-demand
Summary:
The hphpd preset currently just expands /data/users/${USER}/www-hg as the target uri for attaching to hhvm. This works fine on devservers, but the root is different on on-demand instances.

Add fb-specific logic to determine the document root based on the hostname (the same way config.hdf differentiates) and remove the hard-coded value from the preset.

Reviewed By: ebluestein

Differential Revision: D10385895

fbshipit-source-id: 113d097ec5a6d97f0b746c7a67e189f7d2909dac