Move values derived from iterating children into selectors
Summary:
Previously, we derived a bunch of node state while parenting children. This moves those to lazily evaluated selectors instead, using a single selector for the iteration.

The one tricky point here is `isPendingLoad`. There's only one place where this is ever set to `true` and, once it's set to `false`, it's false forever. Unfortunately, that non-pending state was derived from the children + things we've converted to selectors and redux doesn't really have a way to store derived state like that. I *think* that we can get the same result by toggling it when `isLoading` goes from `true` -> `false`, but I'm not 100% sure. The myriad loading state properties (`isLoading`, `isPendingLoad`, `wasFetched`) aren't really documented and it's not clear how they relate. ):

I think `shouldBeShown` and the rest of the MemoizedFieldDeriver's properties can be moved to selectors now.

Reviewed By: captbaritone

Differential Revision: D10355030

fbshipit-source-id: 31f0d3f25d2cbf28653e01154bb8056484bf0dfb