Rename `ReactPerfMonitor` class and sample its values
Summary:
Rather than not reporting values less than 7ms, sample all values the same way we do react profiler timings. Like react profiler timings, always ensure that times >100ms are always reported with a sample size of 1.

This also renames `ReactPerfMonitor` to `ReactDevPerfMonitor` to highlight the fact that it is only used in dev. **This does not change the name of the event tracked**.

Reviewed By: matthewwithanm

Differential Revision: D10412683

fbshipit-source-id: 7d16435ced6138607613c0bb30d30c4754d55d6c