Fix button tooltip
Summary:
This fixes an issue with D10276808. It turns out that the function returned from `addTooltip()` closes over the node so we can't simply re-call it.

The reason `addTooltip()` normally works at all is *very* subtle. It relies on the fact that the returned value changes each render, causing it to be invoked with `null` and then the same element in succession each time. The cleanup is then based on the time between those two (it sets a timeout each render).

Reviewed By: Goom11

Differential Revision: D10435657

fbshipit-source-id: b487d6050d0b953fb917e72cc5a5b1a6f21eb6d3