retry on ECONNREFUSED errors
Summary:
Server crashes may not be server crashes after all.

I spent some time following up with several people who were listed as having several server crashes in the last hour, and none of them had experienced any meaningful interruption with Nuclide. So why weren't they blocked? The heartbeat tries again after 10 seconds, and will clear the error if it succeeds, possibly before the user notices.

My hypothesis is that ECONNREFUSED errors are transient, but our code is assuming that just because we get one, the server has crashed (and we notify the user).

If my hypothesis is correct, the fix here is to retry a few times before we actually report the "Server Crashed" error to the user (and subsequently, to our analytics). If I'm right, then many of these errors will go away.

I hijacked Hanson's code that retries on CERT_NOT_YET_VALID errors, and turned it into a slightly more general retry mechanism. When we get ECONNREFUSED (or CERT_NOT_YET_VALID) we stop sending heartbeats at our normal interval, and retry 3 times at 3 second intervals before we actually consider it an error. If it succeeds within the retry period, we never report it to the user, and it never becomes a "Server Crashed" error. Once we get out of our 'retry' loop, we start the heartbeats again as usual.

If this actually reduces errors, we should probably do an exponential-type backoff rather than a fixed interval.

Reviewed By: matthewwithanm

Differential Revision: D10416156

fbshipit-source-id: 8775a60968693b6634c7260567f1ec9ba782bd70