Stringify keys in CompilationDatabase target cache
Summary:
The targetCache is keyed on the [buckRoot, target, extraArgs] array. Because of object equality rules in javascript, if we construct two arrays then even if the contents are the same they will not hit the same key in the cache.

This causes unecessary Buck builds when multiple files of the same target are opened.

Reviewed By: kcaze

Differential Revision: D10469701

fbshipit-source-id: 0fdb934996001d576cc8d59dc957f59c80755a36