Swap remote server with another remote server if available
Summary:
This is probably controversial but the idea here is simple, if I am working on a devserver and had a file open when on commit 'abcd', and with bookshelf you have the option to open this same file on a different machine when you checkout that hash! This is going to be common use case in on demand.

This change to enable this flow tries to swap the current remote connection in place of the connection with which the file was serialized. Example:

- Machine A, commit 'abcd', README.md is on right pane
- Machine B, checkout 'abcd', open serialized version of README.md in the right pane.

On machine B the README.md would look for machine A to deserialize and replace the placeholder with the an actual editable view. This change enables that.

Bookshelf relies on atom to automatically serialize and deserialize the workspace for us and this method enables us to continue using them by simply modifying the deserialization technique.

Reviewed By: matthewwithanm

Differential Revision: D10371155

fbshipit-source-id: 6cc670c116ca5b7f358934967c749488d80984aa