Add status area showing scroll position
Summary:
Add a status bar showing the scroll position of the output area, and if the bottom has been reached. Also, show if there is more output if output has arrived while the user is scrolled back in the output.

Added Ctrl+HOME and Ctrl+END handlers to jump to the top and bottom of the output.

Reviewed By: ebluestein

Differential Revision: D10508972

fbshipit-source-id: 8fcb66ed17059f5f87198aeb960f3c223ff475a6