Store unsanitized arguments in Buck Toolbar
Summary:
The unsanitized custom arguments that a user inputs into the Buck Toolbar will be stored along with the sanitized custom arguments.

When a user inputs invalid custom arguments in the Nuclide Buck Toolbar, an error will be shown for each of the invalid arguments (i.e. build, run, compile db), letting the user know that the invalid custom arguments will be ignored.

Reviewed By: stepanhruda

Differential Revision: D10415638

fbshipit-source-id: 81f2f8418163d5eff8f4079ad1f4b0a6ea67e6f6