Encode the stall's duration into the performance marker
Summary: Since there is no current way of supplying metadata with a marker (though this is coming in a new proposal), encode the duration of the stall into the marker label itself. We can then extract this in dataswarm and show the length of stalls that took place during a particular trace.

Reviewed By: vdjeric

Differential Revision: D10844148

fbshipit-source-id: a090f042d8ef8a58675d073f2575926b2b3dfb53