Refactor getCompilationDatabase to return an Observable
Summary:
Functionality should not be changed because the conversion just uses fromPromise/switchMap instead of async/await, and the callsites use .toPromise.

This allows us to add a cancel button to abort long-running compilation database calls.

The tricky part here is to make sure that the switch to Observables still works with the caching mechanism. We can use .publishLast to make sure that the final result is shared among all the subscribers that share the call Observable.

Reviewed By: stepanhruda

Differential Revision: D10501094

fbshipit-source-id: 73c116e347b1ac6944bad3b1ecc770affb7bf61e