Update to RxJS 6 via rxjs-compat
Summary:
So far, this seems like a drop-in replacement. The only significant difference I've found is that when errors are unhandled in an observable, in addition to undoing the subscription, RxJS rethrows synchronous errors happening in response to a `next`. In this case, it actually uncovered an issue with Diff View being broken on startup.

For discussion: should we codemod all instances of `import {Observable} from 'rxjs'` to `import {Observable} from 'rxjs-compat'` so that we don't lie about which package things are from? I think it still makes sense to continue rewriting imports to use the minified, bundled version of rx just as we always have, but be honest about the package. This way, we can gradually transition module-by-module from compat to standard rx.

Reviewed By: matthewwithanm

Differential Revision: D10075840

fbshipit-source-id: 08c90a7e990d9c5c267afcabf3dda797d06e4e44