several bug fixes
Summary:
The node adapter misbehaves if a stack frame is asked for more than once per thread stop. So we need to cache the stack.

The gdb adapter needs to protect against asking for stack frames with a start index higher than the number of frames (MI throws an error) now that we're asking for frames at a non-zero index.

Fixed a parsing bug with command aliases.

Reviewed By: ebluestein

Differential Revision: D10693131

fbshipit-source-id: c71da56f57c23eb2c5c123cad57017141fd89e7b