Remeasure heights after delay and log active themes when measuring root height
Summary:
There is a potential race condition in Atom which would cause `atom.themes.onDidChangeActiveTheme` to trigger before the CSS is actually in effect.

This adds a delayed remeasure to account for that, while also adding some logging to help us understand if this potential race condition is actually the cause of the issues we are seeing.

Either:

1. We are mouting after the initial `atom.themes.onDidChangeActiveThemes` but before the CSS is actually in effect.
2. `atom.themes.onDidChangeActiveThemes` is getting called after the theme is loaded, but before the CSS is actually in effect.
3. Somehow `atom.themes.onDidChangeActiveThemes` is never getting called at all.

Reviewed By: wbinnssmith

Differential Revision: D10499870

fbshipit-source-id: 8f7e3285873b1fe75504bc81518145553d123e1b