Scrollbar styles
Summary:
I know this is an impossible task but I figured I'd take a crack at it…

I'm not a huge fan of the border, so I removed that. I think maybe we could get more creative and use `lighten()` instead of just opacity to bring out the thumb a more.

Reviewed By: sunz7

Differential Revision: D10441899

fbshipit-source-id: 80015c0c859db1738071a6a212efb1a41dfaf868