add flag to check whether we should tunnel or not
Summary:
Big Dig automatically tries to connect to the server over an SSH tunnel if the connection is 'insecure'. It determines if a connection is 'insecure' by whether or not it gets a certificate and key back from the SshHandshake.

We're moving to use the RootCanal certs installed by chef on a user's laptop and devserver instead. In addition to being more secure, it makes authentication in Nuclide password-less, even for people who don't have private key files.

This means we'll no longer be doing the SSH handshake, so we can't use that to determine whether the connection is secure. This diff adds a flag telling Big Dig not to *ever* tunnel since I don't think this feature is actively being used anywhere. If we agree that this feature is not useful moving forward, we can remove the code. For now, it unblocks the rest of this work.

Reviewed By: siegebell

Differential Revision: D12830489

fbshipit-source-id: 550aa5bf2583d2d2e257da530a755d37a92725e8