Clean up empty marker sets if we don't create any markers
Summary: Previously we'd leave behind entries in the map for every possible chunk, making the marker map for an editor scale with the number of lines the editor had. Not only would this waste memory, but it would also make `editorHasMarkers` run longer as it would also be bounded by the length of the document.

Reviewed By: sunz7

Differential Revision: D12833140

fbshipit-source-id: d32b25f64069894e7d7ffe0d5e5c7acde16bfb05