Syntax highlighting for Hack attributes
Summary:
This diff adds support to Hack attribute syntax highlighting.

Prompted by having to work with a lot of ReactiveHack code which uses attributes on function parameters
For example, this ReactiveHack function
   <<__Rx,  Description("This class does nothing")>>
   public static function testRxHack<T>(<<__Mutable>> TaoRequest $r): void {}

which looks like this in nuclide
{F142870419}

Turns out, that even the attributes on the function were being treated as constants or functions and the attributes on parameters were being treated as class names. I figured that we probably want to treat Hack Attributes as their own syntax type, so this diff adds support for that.

I tried to maintain the same syntax highlighting on functions while also supporting attributes on parameters. However, attribute that take arguments will now look like other attributes instead of looking like functions

Final result:

{F142870615}

Reviewed By: matthewwithanm

Differential Revision: D12882549

fbshipit-source-id: 2ab314c470236df24e4589adb2ab11fe289dc9ba