make long polling event driven
Summary:
The method for determining if new data had arrived on a pty was to poll for data at a rapid rate and check if data had arrived in the buffer. This worked but it has been improved upon in this diff.

This diff uses Promises to eliminate the need to continously check a buffer for new data which has two benefits: 1.) reduces latency when fetching new data and 2.) reduces system resources on the server.

This is an implementation change, so use of the pty code is effectively unchanged other than its performance and efficiency.

Also switched default encoding to `utf-8`.

Reviewed By: semmypurewal

Differential Revision: D12924853

fbshipit-source-id: b3411f4bb2d1d15705d4788d20978ad25aa28d1b