add backpressure to pty
Summary:
The pty has methods `pause` and `resume` which can be used to add 'backpressure' to the pty. This is useful because the pty process can output data to the thrift callback handler far faster than it can be returned over a TCP or websocket connection.

For such a situation, this diff defines a maximum amount of data to accumulate (`PTY_PAUSE_THRESHOLD_BYTES`) at which point the pty will be told to pause its output so the buffer can be drained by the client over TCP (`this._pty.pause()`), or whatever transport is being used.

After the buffer has been drained to a low enough level (`PTY_RESUME_THRESHOLD_BYTES`), the pty will be told it can resume outputting data (`this._pty.resume()`).

End user impact:
* CTRL+C (SIGINT) now works as expected -- nearly instantly. Prior to this, the CTRL+C command sequence was part of the buffered pty output which meant if the user pressed CTRL+C, the stream would not end until **all the buffered data was sent**. This could, in effect, be **never** if enough data had been buffered (i.e. `cat`ing a large enough file).

Performance impact:
* The size of the buffer is bound to `BUFFER_LENGTH_BYTES`, which is smaller than its previous value of ~2GB. The process will now only allocate 4MB rather than up to 2GB.

Reviewed By: semmypurewal

Differential Revision: D12928459

fbshipit-source-id: b64c714cbaf472c9cb3bdb730b66f26301682d5b