Raise minimum Atom version to 1.28
Summary:
This allows us to remove handling for old versions of Chrome from, for example, the stall observation code.

Atom 1.28 was the first to ship with Electron 2.0, which was the first to include Node 8, an LTS version of Node. This will allow us to reliably use features only in Node 8.

Atom 1.28 Shipped June 21, 2018, more than four months prior to time of writing. I feel enough time has passed that we can expect users to have upgraded, and we will warn them otherwise.

Reviewed By: captbaritone

Differential Revision: D12873204

fbshipit-source-id: 3158160879d2cbf121c8df5853931f8e243d01f4