start server with rootcanal certificates if flag is set
Summary:
Start up the remote Big Dig server with the appropriate certificates if a flag is set. Note that if you change the flag to true, the client will be unable to connect to the server in this diff. It's fixed in a future diff.

Also, adding @[1611198859:Stepan] as a reviewer since he originally wrote the 'reuse' certificateStrategy of the big dig server.

Reviewed By: ebluestein

Differential Revision: D12906679

fbshipit-source-id: 8e457a2e94bc038bac13d1b9766b8f1bfa14a833