add function to extract all CAs from a laptop
Summary:
This diff creates a function that reads all files in a directory, parses them (using split-ca) and returns an array of CAs for use by the client.

We still need to add the correct directories for Windows & Linux, along with the appropriate logic to run the correct code on the correct platform.

Reviewed By: ebluestein

Differential Revision: D12906678

fbshipit-source-id: 15d63f2189a065f50126c1167a884770e44e3bce