Integrate change of revision id with diff service
Summary:
While in split diff mode, the user is able to choose a revision id to compare to. This diff allows the diff service to cache the old version of a focused file, based on the revision selected, and use it to compare to the new version.

Because the diff view navigator is not always visible, changing revision will affect regular diff gutters, they will still be compared to head.

Reviewed By: tjfryan

Differential Revision: D12953270

fbshipit-source-id: c089fa93bc8d316e4e2e8a12735bdf15bda2aa07