Fix too-large root nodes in file tree
Summary:
Previously, a theme change would cause a "forced" remeasure. However, if we happened to not have a root ref at that time, we would never remeasure.

With this commit, we instead *invalidate* when the theme changes. If we don't have a root ref, we'll still be invalidated when we get one, and a new measurement will be taken.

Reviewed By: wbinnssmith

Differential Revision: D13015444

fbshipit-source-id: a1e577b4675a171de6ab496486957f9672ab14c0