Don't break into threads except on breakpoint
Summary:
Entering and staying in all-threads-stopped mode in hhvm has unpleasant side effects with server CLI scripts (they block until the debugger starts threads again). So try to not stay in that state if the user is just evaluating expressions.

Also add a "stopped"/"running" indicator in the status bar.

Reviewed By: ebluestein

Differential Revision: D12977657

fbshipit-source-id: 0c4fe7f265a325e69fadef91fd4c2a92cdc4540a