Make list command more useful
Summary:
With no arguments, the list command on first execution will show code around the current stack frame. Another execution will show the next page of code.

It attempts to reset itself when the stack frame changes, but that didn't work very well. Took out the old code (and changed it from a callback to the debugger emitting events more in line with the rest of our code) and reset the command on every thread and stack frame switch, include if the currently selected frame changes due to a step.

Reviewed By: ebluestein

Differential Revision: D13005241

fbshipit-source-id: a17e8b77ecfec97571fa966ae1bf59c7af8d53c7