fix: source map not properly discovered when attaching to another user's process
Summary:
When attaching on linux, the debugger launch code uses procfs to discover the process binary. The symlink to the binary in procfs isn't world readable, so we fail to resolve the path and thus the buck root. The code that does this is in the setup path and thus doesn't get elevated as the adapter itself does.

Change the discovery code to run as root if it fails to read as the current user.

Reviewed By: ebluestein

Differential Revision: D13000652

fbshipit-source-id: 7d9683864adf1fb22ed55527cba0173d836a9c9c