when using RootCanal certs, request the client cert and verify the user
Summary:
In this diff we reject unauthorized certs, request a peer cert from the client, and check the identity of the user from that cert (and make sure they are authorized to access the devserver).

This also requires us to pipe the useRootCanalCerts flag all the way to the server so that it knows to do these additional checks.

Reviewed By: ebluestein

Differential Revision: D13010688

fbshipit-source-id: 5e34ac7020620fe2b6f54d476f39bc3d09cc3b1c