Use root canal certs on the client if flag is set
Summary:
This sets up a MacOS client to use its local rootcanal certs if a flag is set. One odd thing here is that the certificate also acts as the client's key. I'm not exactly sure how/why that's the case but it's required once we start rejecting unauthorized access (in the next diff).

At this point, we're still using the SshHandshake to start the server (so a password is still required). We'll need to fix that issue under a separate stack of diffs. That's the last piece of the puzzle, and I'll start working on that now.

Reviewed By: ebluestein

Differential Revision: D12830488

fbshipit-source-id: c7f60787b3d92f7e1ea7ca6c848f348a60d0766f