add server implementation
Summary:
Add implementation code to run `observeProcess` and handle client requests. This API must implement all of the methods defined in `ProcessWatcher.thrift` (in the first diff).

The API is small and the general flow is as follows. Note a subsequent diff abstracts these operations into a single function returning an Observable, so end users will not have to perform this flow.
* call `watchProcess` and obtain the id of the process being watched
* retrieve output from the process with `nextMessages`
* unsubscribe/kill the process with `unsubscribe`.

Reviewed By: kcaze

Differential Revision: D13032198

fbshipit-source-id: 7eea0562723b15b1877605a5f3e0fb80270624c7