Replace LazyNestedValueComponent - Implement DAP ErrorResponse and surface SetVariable errors - 9/N
Summary: This diff implements the ErrorResponse mssage specified by the debug adapter protocol so that we can surface errors to the user if the new scopes component fails to set a variable's value. Currently the debug adapter's error message is swallowed and the user just sees the variable hasn't changed, but no indication as to why.

Reviewed By: Goom11

Differential Revision: D10492159

fbshipit-source-id: a22c3934af01c257492f5ec1cfa30bdbebc9f80f