Hard code version of event-stream in node debugger to pre-hacked version
Summary:
vscode-node-debug2 references a version of event-stream that has an injection hack. This is only in devDependencies, and event-stream doesn't make it out to release nuclide, but hard code the version of the module to the pre-hack value for now so it no longer shows up in yarn.lock. See

https://fb.facebook.com/groups/131549423587801/permalink/2049499911792733/

Reviewed By: wbinnssmith

Differential Revision: D13197309

fbshipit-source-id: a0f585260601d0db9a9eb622ac80cb234598ed7e