Change opt-out to require a reason why
Summary: It's too easy to opt out and data shows there's a pattern of people doing it without even trying the new debugger. So make them give a reason and explicitly state that the opt out is temporary. Now they must give a debugger command (although the opt out mechanism is the same)

Reviewed By: tjfryan

Differential Revision: D13138355

fbshipit-source-id: 57db1af928f367026af6f5bb3e8a7531eac65006