Don't retain references to editors
Summary:
Since scrollbars uses an `Immutable.Map`, it's quite easy to be the only thing holding a reference to an editor if they're not properly cleaned up.

Instead, clean up by adding disposables that remove the editor from the map when the editor is destroyed.

This alone is not enough, however. Because the diagnostics scrollbar provider did not properly unsubscribe from diagnostic updates once its editor was destroyed, it was possible to receive diagnostic updates for a destroyed editor, thus causing it to enter the map never to leave.

Amend this by both fixing the diagnostic scrollbar provider *and* by making sure we never add a destroyed editor to the map.

Reviewed By: captbaritone

Differential Revision: D13146020

fbshipit-source-id: 0497b3d6fab6a25cadcb6b63252869cf45c63aec