Fix race condition in release-generate-proxies
Summary:
release-generate-proxies often fails on devservers with multiple CPU's. This is due to a race condition: a process's stdout may or may not be flushed and closed at the time the 'exit' event is generated. If 'exit' happens first, then we don't get all the JSON from the child and parsing fails.

Change the code to ensure that the child process exited successfully AND the stream has pushed all data before we proceed.

Reviewed By: ebluestein

Differential Revision: D13141853

fbshipit-source-id: 7e9c19c21ac029e7971bf40c30e76ef88aacbbe6