parse out the remote server command and pass the command and flags separately
Summary: It's safer to build the actual command on the server, rather than building it on the client and then having the server execute whatever the client sends. The bootstrap service expects the command, the flags, and the arguments to be passed separately, so we send those pieces separately to the SshHandshake.

Reviewed By: cs01

Differential Revision: D13115495

fbshipit-source-id: e349cdc28dd3afacbdffe8df9fcdc3dcfb509187