Fix for syntax highlighting bug for Hack Attributes
Summary:
Ok... I'm convinced writing a syntax highlighter using regex is hard, but I think this might fully support Hack attributes.
The crux of the issue was that we have overlapping definitions for certain syntax. For example, a `<` could be a part of
- an Hack attribute
- a comparison operator
- a bitwise operator
- a heredoc

In fact, before I started working on Hack attributes, Hack attributes, comparison operators and bitwise operators all mapped to the comparison operator syntax highlighting rule.

 **The Hack Attribute Story**

With the first diff, D12882549, we changed this so heredocs, bitwise operators, and hack attributes all mapped to the new hack attribute rule.

With the first fix, D12935138, we fixed this for heredocs. After this, bitwise operators and hack attributes mapped to the same rule. The other 2 mapped to their own.

Finally, this diff fixes the overlap between bitwise operators and hack attributes. The token for the start of the hack attribute and the left-bit-shift operator are identical. To differential between the 2, we need to use additional context. For this, we can just check to see if the line is exclusively whitespace before the `<<`. For Hack attributes on classes and functions, this should be true. For bitwise operators, this should be false (since it's an infix operator). We have to further limit this context to just cases where the Hack attribute is not in a function argument. We use a wrapping pattern to accomplish this.

This is not perfect. For example, if there is a comment on the same line, but before the attributes, we won't highlight as attributes. However, this should cover most cases, and the fallback is just the behavior we had before.

Finally, if we think this is too risky, we can just limit the attribute scope to function arguments

Reviewed By: kcaze

Differential Revision: D13062050

fbshipit-source-id: fd55776a2b107347327d8d01a633ca46a18c6b63